casa_familia_project/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
faker
drf-yasg